app/components/agents/deploiment.py
from vertexai import agent_engines
from app.config.settings import settings

import vertexai

client = vertexai.Client(  # For service interactions via client.agent_engines
    project=settings.GOOGLE_CLOUD_PROJECT,
    location=settings.GOOGLE_CLOUD_LOCATION,
)